Have the uap version of RemoteExecutorTestBase.RemoteInvoke wait in a background thread
Fixes #21275
