+++ src/System.IO.Pipes/tests/AnonymousPipeTests/AnonymousPipeTest.CrossProcess.cs
@@ -10,7 +10,7 @@ namespace System.IO.Pipes.Tests
     public class AnonymousPipeTest_CrossProcess : RemoteExecutorTestBase
     {
         [Fact]
-        [ActiveIssue("https://github.com/dotnet/corefx/issues/21275", TargetFrameworkMonikers.Uap)]
+        [ActiveIssue("https://github.com/dotnet/corefx/issues/21777", TargetFrameworkMonikers.Uap)]
         public void PingPong()
         {
             // Create two anonymous pipes, one for each direction of communication.
