+++ src/System.Threading/tests/SemaphoreTests.cs
@@ -280,7 +280,6 @@ namespace System.Threading.Tests
 
         [Fact]
         [PlatformSpecific(TestPlatforms.Windows)] // names aren't supported on Unix
-        [ActiveIssue(21275, TargetFrameworkMonikers.Uap)]
         public void PingPong()
         {
             // Create names for the two semaphores
